requirements.txt
asgiref==3.4.1
Django==3.2.16
django-filter==21.1
djangorestframework==3.14.0
python-decouple==3.6
pytz==2022.7
sqlparse==0.4.3
typing_extensions==4.1.1
tzdata==2022.7
